+++ src/OpenTelemetry.Exporter.OpenTelemetryProtocol/CHANGELOG.md
@@ -8,7 +8,7 @@
 
 * Added overloads which accept a name to the `TracerProviderBuilder`
   `AddOtlpExporter` extension to allow for more fine-grained options management
-  ([#XXXX](https://github.com/open-telemetry/opentelemetry-dotnet/pull/XXXX))
+  ([#3653](https://github.com/open-telemetry/opentelemetry-dotnet/pull/3653))
 
 ## 1.4.0-alpha.2