+++ src/OpenTelemetry.Exporter.OpenTelemetryProtocol/CHANGELOG.md
@@ -6,6 +6,10 @@
   `AddOtlpExporter` extension to allow for more fine-grained options management
   ([#3648](https://github.com/open-telemetry/opentelemetry-dotnet/pull/3648))
 
+* Added overloads which accept a name to the `TracerProviderBuilder`
+  `AddOtlpExporter` extension to allow for more fine-grained options management
+  ([#3653](https://github.com/open-telemetry/opentelemetry-dotnet/pull/3653))
+
 ## 1.4.0-alpha.2
 
 Released 2022-Aug-18